requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
ConfigArgParse==1.7
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.15
diffusers==0.32.0.dev0
dill==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
fabric==3.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.2.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.18
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.3
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
invoke==2.2.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lightning==2.3.3
lightning-fabric==2.4.0
lightning-utilities==0.11.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.23.0
networkx==3.0
numpy==1.23.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.13
packaging==24.2
pandas==2.0.3
paramiko==3.5.0
peft==0.13.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
ruff==0.8.5
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.0
-e git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf#egg=segment_anything&subdirectory=../../../../ml/SegmentAnything/segment-anything
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
supervision==0.25.0
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
timm==1.0.12
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.12.0
torch==2.4.1+cu118
torch-fidelity==0.3.0
torchaudio==2.4.1+cu118
torchmetrics==1.5.2
torchsearchsorted==1.1
torchvision==0.19.1+cu118
tqdm==4.67.0
transformers==4.46.3
triton==3.0.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
websockets==11.0.3
Werkzeug==3.0.6
wrapt==1.16.0
xformers==0.0.28.post1
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
